* [Records/Owners] Fixed an issue where some task and ticket records could be left owned by a previously deleted worker.

+++ features/cerberusweb.core/api/dao/worker.php
@@ -692,6 +692,16 @@ class DAO_Worker extends Cerb_ORMHelper {
 		
 		$db = DevblocksPlatform::getDatabaseService();
 		
+		// Clear their task assignments
+		$sql = sprintf("UPDATE task SET owner_id = 0 WHERE owner_id = %d", $id);
+		if(false == ($db->ExecuteMaster($sql)))
+			return false;
+		
+		// Clear their ticket assignments
+		$sql = sprintf("UPDATE ticket SET owner_id = 0 WHERE owner_id = %d", $id);
+		if(false == ($db->ExecuteMaster($sql)))
+			return false;
+		
 		$sql = sprintf("DELETE FROM worker WHERE id = %d", $id);
 		if(false == ($db->ExecuteMaster($sql)))
 			return false;
